Add the gitloggraph library and gitree tool.
gitree is a tool I wrote while working at Google that helps me visualize the state of my git repositories. I am now releasing it as an open-source project so that others may use it.

gitloggraph is a library that computes the arguments that should be passed to `git log` in order to view the relationships between a specific set of commits. It is useful for building tools like gitree, and may be used by other tools.