Remove the --username flag, replace it with a new branch selection algorithm.
Previously, gitxl included all branches from remotes whose URL contains the given username. The new algorithm shows all branches whose name matches a local branch. This is simpler to configure, and less finicky. The downside is that to show a remote branch you need to create a local tracking branch, but I haven't found that to be a significant issue.

I also increased the version number to 2.0 because removing a command line flag is a breaking change.