Rewrite git-tree to use `git` directly rather than `libgit`.
`libgit` does not support partial clones, which I need to use now. This algorithm is somewhat simpler, and won't work for everything, but should be Good Enough for my work. The main limitation is that it does not exclude commits that are ancestors of interesting commits but not descended from the merge bases (e.g. if a parentless commit is merged into the repository).