Store Git IDs in binary.
I'm not longer sure this is right. Every single ID we parse gets spit back out as a command line argument EXCEPT the IDs from the straggler-elimination traversal, but those all live in a HashSet that is constantly searched against so converting to binary and back may not be a win anyway.